+++ appenders/src/main/java/com/kdgregory/log4j/aws/internal/kinesis/KinesisLogWriter.java
@@ -74,10 +74,15 @@ extends AbstractLogWriter
     @Override
     protected void createAWSClient()
     {
-        client = tryClientFactory(config.clientFactoryMethod, AmazonKinesis.class);
+        client = tryClientFactory(config.clientFactoryMethod, AmazonKinesis.class, true);
+        if ((client == null) && (config.clientEndpoint == null))
+        {
+            client = tryClientFactory("com.amazonaws.services.kinesis.AmazonKinesisClientBuilder.defaultClient", AmazonKinesis.class, false);
+        }
         if (client == null)
         {
-            client = new AmazonKinesisClient();
+            LogLog.debug(getClass().getSimpleName() + ": creating service client via constructor");
+            client = tryConfigureEndpointOrRegion(new AmazonKinesisClient(), config.clientEndpoint);
         }
     }